Added sample images support

+++ src/flask_app/controller/cmate.py
@@ -7,10 +7,16 @@ import sys
 sys.path.insert(1, str(Path(__file__).resolve().parent.parent.parent/'cmate'))
 from cmate_main import CMate
 
-def blend_images(profile_img, source_img, images_dir, dest_dir):
+def blend_images(profile_img, source_img, profile_dir, source_dir, dest_dir):
     "apply cloth to profile image and save to result"
-    cloth_blender = CMate(str(Path(images_dir)/source_img),
-                          str(Path(images_dir)/profile_img))
+    # verify file exists
+    if not (Path(profile_dir)/profile_img).exists():
+        raise FileNotFoundError("Profile Image not found.")
+    if not (Path(source_dir)/source_img).exists():
+        raise FileNotFoundError("Source Image not found.")
+    # apply cmate
+    cloth_blender = CMate(str(Path(source_dir)/source_img),
+                          str(Path(profile_dir)/profile_img))
     final_img, errors = cloth_blender.apply_cloth()
     filename = "result-"+datetime.now().strftime("%Y-%m-%d-%H-%M-%S")+".jpg"
     cv2.imwrite(str(Path(dest_dir)/filename), final_img)
